Creating Custome User Remove unused files and update paths

+++ dmsApp/models.py
@@ -7,11 +7,14 @@ from cryptography.fernet import Fernet
 from django.conf import settings
 import base64, os
 from django.dispatch import receiver
+from django.contrib.auth.models import AbstractUser
+from django.utils.translation import gettext_lazy as _
+from django.contrib.auth.models import Group, Permission
 
 # Create your models here.
 
 class Post(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
     title = models.CharField(max_length=250)
     description = models.TextField(blank=True, null=True)
     file_path = models.FileField(upload_to='uploads/',blank=True, null=True)
@@ -26,6 +29,39 @@ class Post(models.Model):
         value = fernet.encrypt(str(self.pk).encode())
         value = base64.urlsafe_b64encode(value).decode()
         return reverse("share-file-id", kwargs={"id": (value)})
+    
+# User permissions
+USER_ROLE_CHOICES = (
+    ('Inputter', 'Inputter'),
+    ('Authorizer', 'Authorizer'),
+)
+
+class CustomUser(AbstractUser):
+    profile_picture = models.ImageField(upload_to='profile_pics/', blank=True, null=True)
+    bio = models.TextField(blank=True, null=True)
+    phone_number = models.CharField(max_length=20, blank=True, null=True)
+    address = models.TextField(blank=True, null=True)
+    user_type = models.CharField(choices=USER_ROLE_CHOICES, default='Inputter', max_length=90)
+
+    groups = models.ManyToManyField(Group, related_name='customuser_set', blank=True, verbose_name=_('groups'))
+    user_permissions = models.ManyToManyField(
+        Permission,
+        related_name='customuser_set',
+        blank=True,
+        verbose_name=_('user permissions'),
+        help_text=_('Specific permissions for this user.'),
+    )
+
+    def save(self, *args, **kwargs):
+        # Delete old profile picture if updating
+        try:
+            old_user = CustomUser.objects.get(pk=self.pk)
+            if old_user.profile_picture and self.profile_picture != old_user.profile_picture:
+                old_user.profile_picture.delete(save=False)
+        except CustomUser.DoesNotExist:
+            pass
+
+        super().save(*args, **kwargs)
 
 @receiver(models.signals.post_delete, sender=Post)
 def auto_delete_file_on_delete(sender, instance, **kwargs):
@@ -33,18 +69,34 @@ def auto_delete_file_on_delete(sender, instance, **kwargs):
         if os.path.isfile(instance.file_path.path):
             os.remove(instance.file_path.path)
 
+# @receiver(models.signals.pre_save, sender=Post)
+# def auto_delete_file_on_change(sender, instance, **kwargs):
+#     if not instance.pk:
+#         return False
+
+#     try:
+#         old_file = sender.objects.get(pk=instance.pk).file_path
+#     except sender.DoesNotExist:
+#         return False
+
+#     new_file = instance.file_path
+#     if not old_file == new_file:
+#         if os.path.isfile(old_file.path):
+#             os.remove(old_file.path)
+           
 @receiver(models.signals.pre_save, sender=Post)
 def auto_delete_file_on_change(sender, instance, **kwargs):
     if not instance.pk:
         return False
 
     try:
-        old_file = sender.objects.get(pk=instance.pk).file_path
+        old_post = sender.objects.get(pk=instance.pk)
+        old_file = old_post.file_path
     except sender.DoesNotExist:
         return False
 
     new_file = instance.file_path
     if not old_file == new_file:
-        if os.path.isfile(old_file.path):
+        if bool(old_file) and os.path.isfile(old_file.path):
             os.remove(old_file.path)
-           
+
